Refactor Public Ips
